qcreport: rename 'illumina.py' to 'illuminaqc.py'.
